Moved owner change route to establishment/actions.js
