+++ server/data/ownership.js
@@ -2,19 +2,19 @@
 
 const db = rfr('server/utils/datastore');
 
+const getEstablishmentDetailsQuery =
+`
+SELECT "EstablishmentID"
+FROM cqc."Establishment"
+WHERE "EstablishmentID" = :subEstId;
+`;
+
 const checkEstablishmentQuery =
 `
 SELECT a."EstablishmentID", a."UserUID", b."IsParent", b."ParentID"
 FROM cqc."User" a JOIN cqc."Establishment" b ON a."EstablishmentID" = b."EstablishmentID"
 WHERE a."IsPrimary" = true and b."IsParent" = false AND b."ParentID" IS NOT NULL
 AND a."EstablishmentID" = :subEstId;
-`;
-
-const getEstablishmentDetailsQuery =
-`
-SELECT "EstablishmentID"
-FROM cqc."Establishment"
-WHERE "EstablishmentID" = :subEstId;
 `;
 
 const checkAlreadyRequestedOwnershipQuery =
@@ -40,15 +40,6 @@ FROM cqc."OwnerChangeRequest"
 WHERE cqc."OwnerChangeRequest"."subEstablishmentID" = :subEstId
 ORDER BY cqc."OwnerChangeRequest"."created" DESC
 LIMIT :limit;
-`;
-
-
-const insertNotificationQuery =
-`
-INSERT INTO
-cqc."Notifications"
-("notificationUid", "type", "typeUid", "recipientUserUid", "isViewed")
-VALUES (:nuid, :type, :typUid, :recipientUserUid, :isViewed);
 `;
 
 exports.getEstablishmentDetails = async(params) =>
@@ -86,18 +77,6 @@ exports.changeOwnershipRequest = async (params) =>
             userUid: params.userUid
           },
           type: db.QueryTypes.INSERT
-    })
-
-exports.insertNewNotification = async (params) =>
-    db.query(insertNotificationQuery, {
-        replacements: {
-            nuid: params.notificationUid,
-            type: "OWNERCHANGE",
-            typUid: params.ownerRequestChangeUid,
-            recipientUserUid: params.recipientUserUid,
-            isViewed: false
-        },
-        type: db.QueryTypes.INSERT
     })
 
 exports.lastOwnershipRequest = async (params) =>
