+++ server/routes/accounts/user.js
@@ -17,7 +17,6 @@ const uuid = require('uuid');
 // all user functionality is encapsulated
 const User = require('../../models/classes/user');
 const notifications = rfr('server/data/notifications');
-const ownership = rfr('server/data/ownership');
 
 // default route
 router.route('/').get(async (req, res) => {
@@ -877,89 +876,6 @@ router.route('/swap/establishment/:id').post(async (req, res) => {
   );
 
   return res.set({'Authorization': 'Bearer ' + token}).status(200).json(response);
-});
-
-// PUT request for ownership change request
-router.use('/establishment/:id/ownershipChange', Authorization.hasAuthorisedEstablishment);
-router.route('/establishment/:id/ownershipChange').post(async (req, res) => {
-  try {
-    if(!Number.isInteger(+req.params.id)){
-      return res.status(400).send({
-        message: 'Sub establishment id must be an integer',
-      });
-    }
-
-    const uuidRegex = /^[0-9A-F]{8}-[0-9A-F]{4}-4[0-9A-F]{3}-[89AB][0-9A-F]{3}-[0-9A-F]{12}$/;
-    if (!uuidRegex.test(req.userUid.toUpperCase())){
-      console.error('Invalid user UUID');
-      return res.status(400).send();
-    }
-
-    const params = {
-        ownerRequestChangeUid: uuid.v4(),
-        userUid: req.userUid, //pull the user's uuid out of JWT
-        subEstablishmentId: req.params.id, // and the id from the url
-        permissionRequest: req.body.permissionRequest
-    };
-
-    if (!uuidRegex.test(params.ownerRequestChangeUid.toUpperCase())){
-      console.error('Invalid owner change request UUID');
-      return res.status(400).send();
-    }
-
-    //get posted sub establishment details
-    let getEstiblishmentDetails = await ownership.getEstablishmentDetails(params);
-    if(!getEstiblishmentDetails.length){
-      return res.status(404).send({
-        message: 'Establishment is not found',
-      });
-    }
-
-    //check post establishment ID has IsParent value true and has some ParentID
-    let checkEstablishmentResult = await ownership.checkEstablishment(params);
-    if(!checkEstablishmentResult.length){
-        return res.status(400).send({
-            message: 'Establishment is not a subsidiary',
-          });
-    }else{
-
-        //check already exists ownership records for posted sub establishment id
-        let checkAlreadyRequestedOwnership = await ownership.checkAlreadyRequestedOwnership(params);
-        if(checkAlreadyRequestedOwnership.length){
-          return res.status(400).send({
-              message: `Ownership is already requested for sub establishment id: ${req.params.id}`,
-          });
-        }
-        //save records
-        if (!uuidRegex.test(checkEstablishmentResult[0].UserUID.toUpperCase())){
-          console.error('Invalid recepient user UUID');
-          return res.status(400).send();
-        }
-
-        params.recipientUserUid = checkEstablishmentResult[0].UserUID;
-        let changeRequestResp = await ownership.changeOwnershipRequest(params);
-        if(!changeRequestResp){
-            return res.status(400).send({
-                message: 'Invalid request',
-            });
-        }else{
-            params.notificationUid = uuid.v4();
-            if (!uuidRegex.test(params.notificationUid.toUpperCase())){
-              console.error('Invalid notification UUID');
-              return res.status(400).send();
-            }
-
-            let addNotificationResp = await ownership.insertNewNotification(params);
-            if(addNotificationResp){
-                let resp = await ownership.lastOwnershipRequest(params);
-                return res.status(201).send(resp[0]);
-            }
-        }
-    }
-  } catch(e) {
-    console.error("/establishment/:id/ownershipChange: ERR: ", err.message);
-    return res.status(503).send({});        // intentionally an empty JSON response
-  }
 });
 
 module.exports = router;
